// src/modules/dblobs/dblobs.types.ts
export type DBlobDBContextId = 'global';
export type DBlobDBScopeId = 'app-chat' | 'app-draw' | 'attachment-drafts';
// ...
export type DBlobDocumentAsset = DBlobAssetImplV1<DBlobAssetType.DOCUMENT, DBlobMimeType.DOCUMENT_PDF | DBlobMimeType.DOCUMENT_XLSX | DBlobMimeType.DOCUMENT_DOCX, DocumentAssetMetadata>;
// type DBlobTextAsset = DBlobAssetImplV1<DBlobAssetType.TEXT, DBlobDataMimeType.DOCUMENT_PLAIN, {}>;


// DB - Asset Generic Type

interface DBlobAssetImplV1<TAssetType extends DBlobAssetType, TMime extends DBlobMimeType, TMeta extends Record<string, any>> {
  id: DBlobAssetId; // Unique identifier
  assetType: TAssetType; // Type of asset, used for discrimination

  label: string; // Textual representation
  data: DBlobAssetData<TMime>; // Original data as a BlobData object
  origin: DBlobAssetOrigin; // Source of the data (e.g., "upload", "generated")

  createdAt: Date; // Creation date
  updatedAt: Date; // Last updated date

  metadata: TMeta; // Flexible metadata for specific .type(s)
  // cache: Record<string, DBlobData<DBlobMimeType>>; // Cached conversions as BlobData objects
  cache: {
    thumb256?: DBlobAssetData<DBlobMimeType.IMG_WEBP | DBlobMimeType.IMG_JPEG>; // Cache for the thumbnail-256 conversion
  };
}

export type DBlobAssetId = string;

export enum DBlobAssetType {
  IMAGE = 'image',
  AUDIO = 'audio',
  // VIDEO = 'video',
  DOCUMENT = 'document',
  // EGO = 'ego',
}


// Asset Data

interface DBlobAssetData<M extends DBlobMimeType> {
  mimeType: M;
  base64: string; // Base64 encoded content (not a data URL)
  // NOTE: the data url will be "data:${mimeType};base64,${base64}"
  // size?: number; // Size in bytes (optional)
  // altMimeType?: DBlobMimeType; // Alternative MIME type for the input (optional)
  // altData?: string; // Alternative data for the input (optional)
}

export enum DBlobMimeType {
  IMG_PNG = 'image/png', IMG_JPEG = 'image/jpeg', IMG_WEBP = 'image/webp',
  AUDIO_MPEG = 'audio/mpeg', AUDIO_WAV = 'audio/wav',
  // VIDEO_MP4 = 'video/mp4',
  DOCUMENT_PDF = 'application/pdf',
  DOCUMENT_XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  DOCUMENT_DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  // DOCUMENT_PLAIN = 'text/plain', DOCUMENT_HTML = 'text/html',
}


// Asset Origin

type DBlobAssetOrigin = UserOrigin | GeneratedOrigin;
// ...
export interface DocumentAssetMetadata {
  pageCount?: number;
  author?: string;
  title?: string;
  subject?: string;
  keywords?: string[];
  description?: string;
}